[SPARK-49426][CONNECT][SQL] Create a shared interface for DataFrameWriterV2
### What changes were proposed in this pull request?
This PR creates a shared interface for DataFrameWriterV2.

### Why are the changes needed?
We are creating a shared Scala Spark SQL interface for Classic and Connect.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #47962 from hvanhovell/SPARK-49426.

Authored-by: Herman van Hovell <herman@databricks.com>
